rename helpers/->lib/ in server; add path aliases for server src and lib
