+++ server/src/lib/auth/is-authorized.ts
