+++ server/src/lib/data/merge-activities-and-relations.ts
